-WALL -WERROR fixes

+++ sources/bltns/bltn_type/bltn_type_free.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   bltn_type_free.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fnancy <fnancy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: hgranule <hgranule@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/24 19:22:17 by fnancy            #+#    #+#             */
-/*   Updated: 2019/09/27 17:55:17 by fnancy           ###   ########.fr       */
+/*   Updated: 2019/11/19 02:01:33 by hgranule         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,9 +28,9 @@ int			bltn_type_freespl(char ***spl)
 
 void		bltn_type_freedstrings(t_darr *res)
 {
-	int		i;
+	size_t		i;
 
-	i = -1;
+	i = (size_t)-1;
 	while (++i < res->count)
 		dstr_del(&res->strings[i]);
 }
